requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
-f https://download.pytorch.org/whl/cu121

google-genai==1.5.0
pydantic==2.10.6
comfy==0.0.1
einops==0.8.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
Imath==0.0.2
jaxtyping==0.2.36
kiui==0.2.10
mathutils==3.3.0
matplotlib==3.9.2
git+https://github.com/NVlabs/nvdiffrast
omegaconf==2.3.0
onnxruntime==1.18.0
open3d==0.18.0
opencv_python==4.10.0.84
opencv_python_headless==4.9.0.80
OpenEXR==3.3.2
Pillow==11.0.0
plyfile==1.1
PyGLM==2.7.1
pygltflib==1.16.3
PyMCubes==0.1.4
pymeshlab==2023.12.post1
PyOpenGL==3.1.0
PyYAML==6.0.1
rembg==2.0.57
scipy==1.13.1
torch-scatter 
-f https://data.pyg.org/whl/torch-2.4.0+cu121.html
tqdm==4.66.4
transformers==4.47.1
trimesh==4.4.0
webdataset==0.2.86
xatlas==0.0.9
ninja
sentencepiece
accelerate
git+https://github.com/huggingface/diffusers
einops
numpy==1.26.3
fvcore
pytorch-lightning==2.2.0
PEFT==0.10.0
pyrender==0.1.45
timm
xformers==0.0.27.post1